refactor: simplify basket work dashboard
